pr #113 comment resolving, refactoring and license
