+++ pom.xml
@@ -199,7 +199,7 @@
                 <version>3.3</version>
                 <configuration>
                     <compilerArgument>-Xlint</compilerArgument>
-		    <source>1.8</source>
+                    <source>1.8</source>
                     <target>1.8</target>
                 </configuration>
             </plugin>
