--- src/main/java/in/zapr/druid/druidry/having/DruidHaving.java
@@ -1,11 +1,0 @@
-package in.zapr.druid.druidry.having;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-@Getter
-@EqualsAndHashCode
-public abstract class DruidHaving {
-    protected String type;
-
-}
