+++ src/main/java/in/zapr/druid/druidry/query/aggregation/DruidGroupByQuery.java
@@ -26,7 +26,7 @@ import in.zapr.druid.druidry.aggregator.DruidAggregator;
 import in.zapr.druid.druidry.dimension.DruidDimension;
 import in.zapr.druid.druidry.filter.DruidFilter;
 import in.zapr.druid.druidry.granularity.Granularity;
-import in.zapr.druid.druidry.having.DruidHaving;
+import in.zapr.druid.druidry.filter.havingSpec.HavingSpec;
 import in.zapr.druid.druidry.limitSpec.DefaultLimitSpec;
 import in.zapr.druid.druidry.postAggregator.DruidPostAggregator;
 import in.zapr.druid.druidry.query.QueryType;
@@ -42,7 +42,7 @@ import lombok.NonNull;
 public class DruidGroupByQuery extends DruidAggregationQuery {
 
     private DefaultLimitSpec limitSpec;
-    private DruidHaving having;
+    private HavingSpec having;
     @NonNull
     private List<DruidDimension> dimensions;
 
@@ -55,7 +55,7 @@ public class DruidGroupByQuery extends DruidAggregationQuery {
                               DruidFilter filter,
                               List<DruidAggregator> aggregators,
                               List<DruidPostAggregator> postAggregators,
-                              DruidHaving having,
+                              HavingSpec having,
                               @NonNull List<Interval> intervals,
                               Context context) {
