+++ src/test/java/in/zapr/druid/druidry/filter/havingSpec/AndHavingTest.java
@@ -1,7 +1,24 @@
-package in.zapr.druid.druidry.having;
+/*
+ * Copyright 2018-present Red Brick Lane Marketing Solutions Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package in.zapr.druid.druidry.filter.havingSpec;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -12,6 +29,7 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
+
 public class AndHavingTest {
 
     private static ObjectMapper objectMapper;
@@ -24,15 +42,15 @@ public class AndHavingTest {
     @Test
     public void testAllFields() throws JSONException, JsonProcessingException {
 
-        DruidHaving equalToHaving1 = new EqualToHaving("Hello", "World");
-        DruidHaving equalToHaving2 = new EqualToHaving("Peace", 15);
+        HavingSpec equalToHaving1 = new EqualToHaving("Hello", 14);
+        HavingSpec equalToHaving2 = new EqualToHaving("Peace", 15);
 
         AndHaving having = new AndHaving(Arrays.asList(equalToHaving1, equalToHaving2));
 
         JSONObject having1 = new JSONObject();
         having1.put("type", "equalTo");
         having1.put("aggregation", "Hello");
-        having1.put("value", "World");
+        having1.put("value", 14);
 
         JSONObject having2 = new JSONObject();
         having2.put("type", "equalTo");
@@ -58,8 +76,8 @@ public class AndHavingTest {
 
     @Test
     public void testEquals() {
-        DruidHaving equalToHaving1 = new EqualToHaving("Hello", "World");
-        DruidHaving equalToHaving2 = new EqualToHaving("Hello", "World");
+        HavingSpec equalToHaving1 = new EqualToHaving("Hello", 15);
+        HavingSpec equalToHaving2 = new EqualToHaving("Hello", 15);
 
         Assert.assertEquals(equalToHaving1, equalToHaving2);
     }
