+++ src/test/java/in/zapr/druid/druidry/filter/havingSpec/EqualToHavingTest.java
@@ -1,7 +1,24 @@
-package in.zapr.druid.druidry.having;
+/*
+ * Copyright 2018-present Red Brick Lane Marketing Solutions Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package in.zapr.druid.druidry.filter.havingSpec;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.skyscreamer.jsonassert.JSONAssert;
@@ -19,7 +36,7 @@ public class EqualToHavingTest {
 
     @Test
     public void testAllFieldsNumeric() throws JSONException, JsonProcessingException {
-        DruidHaving having = new EqualToHaving("count", 2);
+        HavingSpec having = new EqualToHaving("count", 2);
 
         JSONObject jsonObject = new JSONObject();
         jsonObject.put("type", "equalTo");
@@ -31,22 +48,8 @@ public class EqualToHavingTest {
         JSONAssert.assertEquals(expectedJSON, actualJSON, JSONCompareMode.NON_EXTENSIBLE);
     }
 
-    @Test
-    public void testAllFieldsString() throws JSONException, JsonProcessingException {
-        DruidHaving having = new EqualToHaving("count", "2");
-
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put("type", "equalTo");
-        jsonObject.put("aggregation", "count");
-        jsonObject.put("value", "2");
-
-        String actualJSON = objectMapper.writeValueAsString(having);
-        String expectedJSON = jsonObject.toString();
-        JSONAssert.assertEquals(expectedJSON, actualJSON, JSONCompareMode.NON_EXTENSIBLE);
-    }
-
     @Test(expectedExceptions = NullPointerException.class)
     public void testSettingRequiredFieldAsNull() {
-        EqualToHaving having = new EqualToHaving(null, "World");
+        EqualToHaving having = new EqualToHaving(null, 2);
     }
 }
