+++ src/test/java/in/zapr/druid/druidry/filter/havingSpec/FilterHavingTest.java
@@ -1,14 +1,32 @@
-package in.zapr.druid.druidry.having;
+/*
+ * Copyright 2018-present Red Brick Lane Marketing Solutions Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package in.zapr.druid.druidry.filter.havingSpec;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import in.zapr.druid.druidry.filter.SelectorFilter;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
+
+import in.zapr.druid.druidry.filter.SelectorFilter;
 
 public class FilterHavingTest {
     private static ObjectMapper objectMapper;
@@ -20,7 +38,7 @@ public class FilterHavingTest {
 
     @Test
     public void testLongField() throws JSONException, JsonProcessingException {
-        DruidHaving filter = new FilterHaving(new SelectorFilter("Hello", 1488498926000L));
+        HavingSpec filter = new FilterHaving(new SelectorFilter("Hello", 1488498926000L));
         JSONObject filterJsonObject = new JSONObject();
         filterJsonObject.put("type", "selector");
         filterJsonObject.put("dimension", "Hello");
@@ -38,7 +56,7 @@ public class FilterHavingTest {
     @Test(expectedExceptions = NullPointerException.class)
     public void testDimensionMissing() {
         String dimension = null;
-        DruidHaving filter = new FilterHaving(null);
+        HavingSpec filter = new FilterHaving(null);
     }
 }
