+++ src/test/java/in/zapr/druid/druidry/filter/havingSpec/LessThanHavingTest.java
@@ -1,7 +1,24 @@
-package in.zapr.druid.druidry.having;
+/*
+ * Copyright 2018-present Red Brick Lane Marketing Solutions Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package in.zapr.druid.druidry.filter.havingSpec;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.skyscreamer.jsonassert.JSONAssert;
@@ -9,7 +26,7 @@ import org.skyscreamer.jsonassert.JSONCompareMode;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-public class GreaterThanHavingTest {
+public class LessThanHavingTest {
     private static ObjectMapper objectMapper;
 
     @BeforeClass
@@ -19,10 +36,10 @@ public class GreaterThanHavingTest {
 
     @Test
     public void testAllFieldsNumeric() throws JSONException, JsonProcessingException {
-        GreaterThanHaving having = new GreaterThanHaving("count", 2);
+        HavingSpec having = new LessThanHaving("count", 2);
 
         JSONObject jsonObject = new JSONObject();
-        jsonObject.put("type", "greaterThan");
+        jsonObject.put("type", "lessThan");
         jsonObject.put("aggregation", "count");
         jsonObject.put("value", 2);
 
@@ -31,22 +48,8 @@ public class GreaterThanHavingTest {
         JSONAssert.assertEquals(expectedJSON, actualJSON, JSONCompareMode.NON_EXTENSIBLE);
     }
 
-    @Test
-    public void testAllFieldsString() throws JSONException, JsonProcessingException {
-        GreaterThanHaving having = new GreaterThanHaving("count", "2");
-
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put("type", "greaterThan");
-        jsonObject.put("aggregation", "count");
-        jsonObject.put("value", "2");
-
-        String actualJSON = objectMapper.writeValueAsString(having);
-        String expectedJSON = jsonObject.toString();
-        JSONAssert.assertEquals(expectedJSON, actualJSON, JSONCompareMode.NON_EXTENSIBLE);
-    }
-
     @Test(expectedExceptions = NullPointerException.class)
     public void testSettingRequiredFieldAsNull() {
-        GreaterThanHaving having = new GreaterThanHaving(null, "World");
+        LessThanHaving having = new LessThanHaving(null, 2);
     }
 }
