+++ src/test/java/in/zapr/druid/druidry/filter/havingSpec/NotHavingTest.java
@@ -1,7 +1,24 @@
-package in.zapr.druid.druidry.having;
+/*
+ * Copyright 2018-present Red Brick Lane Marketing Solutions Pvt. Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package in.zapr.druid.druidry.filter.havingSpec;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.skyscreamer.jsonassert.JSONAssert;
@@ -21,14 +38,14 @@ public class NotHavingTest {
     @Test
     public void testAllFields() throws JSONException, JsonProcessingException {
 
-        DruidHaving equalToHaving1 = new EqualToHaving("Hello", "World");
+        HavingSpec equalToHaving1 = new EqualToHaving("Hello", 2);
 
-        DruidHaving having = new NotHaving(equalToHaving1);
+        HavingSpec having = new NotHaving(equalToHaving1);
 
         JSONObject havingInner = new JSONObject();
         havingInner.put("type", "equalTo");
         havingInner.put("aggregation", "Hello");
-        havingInner.put("value", "World");
+        havingInner.put("value", 2);
 
         JSONObject jsonObject = new JSONObject();
         jsonObject.put("type", "not");
